refactor(boot): strip legacy SaveFile migration, rename migration.ts → boot.ts
App has not been used for production data, so v1→v2 upgrade, the standalone
bts-rules-v1 / theme keys, and the removed `emoji` field on category nodes
were all dead code. Drop them along with V1SaveFileSchema and the tests that
only exercised those paths. Drift healing and corrupt-blob backup are kept
(not backward-compat). Renamed the file and the entry point to reflect what
remains: boot init, not migration.
